challenge-polyai-smartsort/ml-model/requirements.txt
tensorflow==2.16.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
